+++ lib/rundoc/code_command/file_command/remove.rb
@@ -7,7 +7,10 @@ class Rundoc::CodeCommand::FileCommand
     end
 
     def to_md(env)
-      raise "must call write in its own code section" unless env[:commands].empty?
+      if env[:commands].any? { |c| c[:object].not_hidden? }
+        raise "Must call remove in its own code section"
+      end
+
       env[:before] << "In file `#{filename}` remove:"
       env[:before] << NEWLINE
       nil